Remove TestBenchTestCase
